Bug note
Noted a bug for a log/exp mismatch (was exponentiated base e via tf.exp, but parameters were initialized in log base 2). This does not affect the model's ability to learn, but it would be better if the PKdM were initially made using log base e (or these lines were updated to exponentiate in log base 2).